+++ logfire/_internal/auto_trace/import_hook.py
@@ -1,14 +1,16 @@
 from __future__ import annotations
 
+import ast
 import sys
 from dataclasses import dataclass
 from importlib.abc import Loader, MetaPathFinder
 from importlib.machinery import ModuleSpec
 from importlib.util import spec_from_loader
 from types import ModuleType
-from typing import TYPE_CHECKING, Callable, Iterator, Sequence, cast
+from typing import TYPE_CHECKING, Any, Callable, Iterator, Sequence, cast
 
-from .rewrite_ast import exec_source
+from ..utils import log_internal_error
+from .rewrite_ast import compile_source
 from .types import AutoTraceModule
 
 if TYPE_CHECKING:
@@ -48,10 +50,38 @@ class LogfireFinder(MetaPathFinder):
             if not source:
                 continue
 
-            if not self.modules_filter(AutoTraceModule(fullname, plain_spec.origin)):
+            # We fully expect plain_spec.origin and self.get_filename(...)
+            # to be the same thing (a valid filename), but they're optional.
+            filename = plain_spec.origin
+            if not filename:  # pragma: no cover
+                try:
+                    filename = cast('str | None', plain_spec.loader.get_filename(fullname))  # type: ignore
+                except Exception:
+                    pass
+
+            if not self.modules_filter(AutoTraceModule(fullname, filename)):
                 return None  # tell the import system to try the next meta path finder
 
-            loader = LogfireLoader(plain_spec, source, self.logfire, self.min_duration)
+            try:
+                tree = ast.parse(source)
+            except Exception:  # pragma: no cover
+                # The plain finder gave us invalid source code. Try another one.
+                # A very likely case is that the source code really is invalid,
+                # in which case we'll eventually return None and the normal system will raise the error,
+                # giving the user a normal traceback instead of a confusing and ugly one mentioning logfire.
+                continue
+
+            filename = filename or f'<{fullname}>'
+
+            try:
+                execute = compile_source(tree, filename, fullname, self.logfire, self.min_duration)
+            except Exception:  # pragma: no cover
+                # Auto-tracing failed with an unexpected error. Ensure that this doesn't crash the whole application.
+                # This error handling is why we compile in the finder. Once we return a loader, we've committed to it.
+                log_internal_error()
+                return None  # tell the import system to try the next meta path finder
+
+            loader = LogfireLoader(plain_spec, execute)
             return spec_from_loader(fullname, loader)
 
     def _find_plain_specs(
@@ -79,29 +109,15 @@ class LogfireLoader(Loader):
     plain_spec: ModuleSpec
     """A spec for the module that was returned by another meta path finder (see `LogfireFinder._find_plain_specs`)."""
 
-    source: str
-    """The source code of the module, as returned by `plain_spec.loader.get_source(fullname)`."""
-
-    logfire: Logfire
-    min_duration: int
+    execute: Callable[[dict[str, Any]], None]
+    """A function which accepts module globals and executes the compiled code."""
 
     def exec_module(self, module: ModuleType):
         """Execute a modified AST of the module's source code in the module's namespace.
 
         This is called by the import system.
         """
-        # We fully expect self.plain_spec.origin, module.__file__, and self.get_filename(...)
-        # to all be the same thing (a valid filename), but technically they're all optional,
-        # so this is just an abundance of caution.
-        filename = self.plain_spec.origin or module.__file__
-        if not filename:  # pragma: no cover
-            try:
-                filename = self.get_filename(module.__name__)
-            except Exception:
-                pass
-        filename = filename or f'<{module.__name__}>'
-
-        exec_source(self.source, filename, module.__name__, module.__dict__, self.logfire, self.min_duration)
+        self.execute(module.__dict__)
 
     # This is required when `exec_module` is defined.
     # It returns None to indicate that the usual module creation process should be used.