+++ logfire/_internal/main.py
@@ -8,14 +8,26 @@ import traceback
 import warnings
 from functools import cached_property
 from time import time
-from typing import TYPE_CHECKING, Any, Callable, ContextManager, Iterable, Literal, Sequence, TypeVar, Union, cast
+from typing import (
+    TYPE_CHECKING,
+    Any,
+    Callable,
+    ContextManager,
+    Iterable,
+    Literal,
+    Sequence,
+    TypeVar,
+    Union,
+    cast,
+    overload,
+)
 
 import opentelemetry.context as context_api
 import opentelemetry.trace as trace_api
 from opentelemetry.metrics import CallbackT, Counter, Histogram, UpDownCounter
 from opentelemetry.sdk.trace import ReadableSpan, Span
 from opentelemetry.semconv.trace import SpanAttributes
-from opentelemetry.trace import StatusCode, Tracer
+from opentelemetry.trace import SpanContext, StatusCode, Tracer
 from opentelemetry.util import types as otel_types
 from typing_extensions import LiteralString, ParamSpec
 
@@ -154,6 +166,7 @@ class Logfire:
         _tags: Sequence[str] | None = None,
         _span_name: str | None = None,
         _level: LevelName | int | None = None,
+        _links: Sequence[tuple[SpanContext, otel_types.Attributes]] = (),
     ) -> LogfireSpan:
         try:
             stack_info = get_user_stack_info()
@@ -200,6 +213,7 @@ class Logfire:
                 otlp_attributes,
                 self._spans_tracer,
                 json_schema_properties,
+                links=_links,
             )
         except Exception:
             log_internal_error()
@@ -228,7 +242,7 @@ class Logfire:
         try:
             msg_template: str = attributes[ATTRIBUTES_MESSAGE_TEMPLATE_KEY]  # type: ignore
             attributes[ATTRIBUTES_MESSAGE_KEY] = logfire_format(msg_template, function_args, self._config.scrubber)
-            if json_schema_properties := attributes_json_schema_properties(function_args):
+            if json_schema_properties := attributes_json_schema_properties(function_args):  # pragma: no branch
                 attributes[ATTRIBUTES_JSON_SCHEMA_KEY] = attributes_json_schema(json_schema_properties)
             attributes.update(user_attributes(function_args))
             return self._fast_span(name, attributes)
@@ -492,6 +506,7 @@ class Logfire:
         _tags: Sequence[str] | None = None,
         _span_name: str | None = None,
         _level: LevelName | None = None,
+        _links: Sequence[tuple[SpanContext, otel_types.Attributes]] = (),
         **attributes: Any,
     ) -> LogfireSpan:
         """Context manager for creating a span.
@@ -510,6 +525,7 @@ class Logfire:
             _span_name: The span name. If not provided, the `msg_template` will be used.
             _tags: An optional sequence of tags to include in the span.
             _level: An optional log level name.
+            _links: An optional sequence of links to other spans. Each link is a tuple of a span context and attributes.
             attributes: The arguments to include in the span and format the message template with.
                 Attributes starting with an underscore are not allowed.
         """
@@ -521,14 +537,16 @@ class Logfire:
             _tags=_tags,
             _span_name=_span_name,
             _level=_level,
+            _links=_links,
         )
 
+    @overload
     def instrument(
         self,
         msg_template: LiteralString | None = None,
         *,
         span_name: str | None = None,
-        extract_args: bool = True,
+        extract_args: bool | Iterable[str] = True,
         allow_generator: bool = False,
     ) -> Callable[[Callable[P, R]], Callable[P, R]]:
         """Decorator for instrumenting a function as a span.
@@ -547,10 +565,59 @@ class Logfire:
         Args:
             msg_template: The template for the span message. If not provided, the module and function name will be used.
             span_name: The span name. If not provided, the `msg_template` will be used.
-            extract_args: Whether to extract arguments from the function signature and log them as span attributes.
+            extract_args: By default, all function call arguments are logged as span attributes.
+                Set to `False` to disable this, or pass an iterable of argument names to include.
             allow_generator: Set to `True` to prevent a warning when instrumenting a generator function.
                 Read https://logfire.pydantic.dev/docs/guides/advanced/generators/#using-logfireinstrument first.
         """
+
+    @overload
+    def instrument(self, func: Callable[P, R]) -> Callable[P, R]:
+        """Decorator for instrumenting a function as a span, with default configuration.
+
+        ```py
+        import logfire
+
+        logfire.configure()
+
+
+        @logfire.instrument
+        def my_function(a: int):
+            logfire.info('new log {a=}', a=a)
+        ```
+        """
+
+    def instrument(  # type: ignore[reportInconsistentOverload]
+        self,
+        msg_template: Callable[P, R] | LiteralString | None = None,
+        *,
+        span_name: str | None = None,
+        extract_args: bool | Iterable[str] = True,
+        allow_generator: bool = False,
+    ) -> Callable[[Callable[P, R]], Callable[P, R]] | Callable[P, R]:
+        """Decorator for instrumenting a function as a span.
+
+        ```py
+        import logfire
+
+        logfire.configure()
+
+
+        @logfire.instrument('This is a span {a=}')
+        def my_function(a: int):
+            logfire.info('new log {a=}', a=a)
+        ```
+
+        Args:
+            msg_template: The template for the span message. If not provided, the module and function name will be used.
+            span_name: The span name. If not provided, the `msg_template` will be used.
+            extract_args: By default, all function call arguments are logged as span attributes.
+                Set to `False` to disable this, or pass an iterable of argument names to include.
+            allow_generator: Set to `True` to prevent a warning when instrumenting a generator function.
+                Read https://logfire.pydantic.dev/docs/guides/advanced/generators/#using-logfireinstrument first.
+        """
+        if callable(msg_template):
+            return self.instrument()(msg_template)
         return instrument(self, tuple(self._tags), msg_template, span_name, extract_args, allow_generator)
 
     def log(
@@ -1751,11 +1818,13 @@ class LogfireSpan(ReadableSpan):
         otlp_attributes: dict[str, otel_types.AttributeValue],
         tracer: Tracer,
         json_schema_properties: JsonSchemaProperties,
+        links: Sequence[tuple[SpanContext, otel_types.Attributes]],
     ) -> None:
         self._span_name = span_name
         self._otlp_attributes = otlp_attributes
         self._tracer = tracer
         self._json_schema_properties = json_schema_properties
+        self._links = list(trace_api.Link(context=context, attributes=attributes) for context, attributes in links)
 
         self._added_attributes = False
         self._end_on_exit: bool | None = None
@@ -1775,6 +1844,7 @@ class LogfireSpan(ReadableSpan):
                 self._span = self._tracer.start_span(
                     name=self._span_name,
                     attributes=self._otlp_attributes,
+                    links=self._links,
                 )
             if self._token is None:  # pragma: no branch
                 self._token = context_api.attach(trace_api.set_span_in_context(self._span))
@@ -1862,6 +1932,12 @@ class LogfireSpan(ReadableSpan):
         """Sets the given attributes on the span."""
         for key, value in attributes.items():
             self.set_attribute(key, value)
+
+    def add_link(self, context: SpanContext, attributes: otel_types.Attributes = None) -> None:
+        if self._span is None:
+            self._links += [trace_api.Link(context=context, attributes=attributes)]
+        else:
+            self._span.add_link(context, attributes)
 
     # TODO(Marcelo): We should add a test for `record_exception`.
     def record_exception(