+++ logfire/_internal/metrics.py
@@ -6,7 +6,6 @@ from threading import Lock
 from typing import Any, Generic, Sequence, TypeVar
 from weakref import WeakSet
 
-from opentelemetry.context import Context
 from opentelemetry.metrics import (
     CallbackT,
     Counter,
@@ -255,9 +254,12 @@ class _ProxyCounter(_ProxyInstrument[Counter], Counter):
         self,
         amount: int | float,
         attributes: Attributes | None = None,
-        context: Context | None = None,
+        # Starting with opentelemetry-sdk 1.28.0, these methods accept an additional optional `context` argument.
+        # This is passed to the underlying instrument using `*args, **kwargs` for compatibility with older versions.
+        *args: Any,
+        **kwargs: Any,
     ) -> None:
-        self._instrument.add(amount, attributes, context)
+        self._instrument.add(amount, attributes, *args, **kwargs)
 
     def _create_real_instrument(self, meter: Meter) -> Counter:
         return meter.create_counter(self._name, self._unit, self._description)
@@ -268,9 +270,10 @@ class _ProxyHistogram(_ProxyInstrument[Histogram], Histogram):
         self,
         amount: int | float,
         attributes: Attributes | None = None,
-        context: Context | None = None,
+        *args: Any,
+        **kwargs: Any,
     ) -> None:
-        self._instrument.record(amount, attributes, context)
+        self._instrument.record(amount, attributes, *args, **kwargs)
 
     def _create_real_instrument(self, meter: Meter) -> Histogram:
         return meter.create_histogram(self._name, self._unit, self._description)
@@ -302,9 +305,10 @@ class _ProxyUpDownCounter(_ProxyInstrument[UpDownCounter], UpDownCounter):
         self,
         amount: int | float,
         attributes: Attributes | None = None,
-        context: Context | None = None,
+        *args: Any,
+        **kwargs: Any,
     ) -> None:
-        self._instrument.add(amount, attributes, context)
+        self._instrument.add(amount, attributes, *args, **kwargs)
 
     def _create_real_instrument(self, meter: Meter) -> UpDownCounter:
         return meter.create_up_down_counter(self._name, self._unit, self._description)
@@ -317,9 +321,10 @@ if Gauge is not None:  # pragma: no branch
             self,
             amount: int | float,
             attributes: Attributes | None = None,
-            context: Context | None = None,
+            *args: Any,
+            **kwargs: Any,
         ) -> None:  # pragma: no cover
-            self._instrument.set(amount, attributes, context)
+            self._instrument.set(amount, attributes, *args, **kwargs)
 
         def _create_real_instrument(self, meter: Meter):  # pragma: no cover
             return meter.create_gauge(self._name, self._unit, self._description)